+++ town-api/src/main/java/entity/UpdateInfoDO.java
@@ -7,12 +7,12 @@ import java.io.Serializable;
 @Data
 public class UpdateInfoDO implements Serializable {
     private Integer updateId;
-    private Integer infoId;
+    private Long infoId;
     private Integer infoType;
     private byte[] beforeMsg;
     private byte[] afterMsg;
     private Long updateTime;
-    private Integer updateUserTel;
+    private Long updateUserTel;
     private String updateName;
 }
